pyproject.toml
[tool.poetry]
name = "msys2-autobuild-controller"
version = "0.1.0"
description = ""
authors = [""]

[tool.poetry.dependencies]
python = "^3.8.1"
Flask = "^2.0.0"
requests = "^2.25.1"
cryptography = ">=3.4.7"
PyGithub = "^1.54.1"
gunicorn = "^20.1.0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
flake8 = "^6.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
